+++ scripts/Publish-Release.ps1
@@ -6,7 +6,7 @@ param(
 
     [string]$ReferenceBundleDir,
 
-    [string]$ReleaseNotesPath = 'docs\RELEASE-NOTES-0.0.134.md',
+    [string]$ReleaseNotesPath = 'docs\RELEASE-NOTES-0.0.135.md',
 
     [string]$ReleaseBranch = 'master',